requirements.txt
asgiref==3.3.1
Django==3.1.7
django-environ==0.4.5
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
flake8==3.9.0
gunicorn==20.0.4
isort==5.7.0
mccabe==0.6.1
Pillow==8.1.2
psycopg2-binary==2.8.5
pycodestyle==2.7.0
pyflakes==2.3.0
pytz==2021.1
sorl-thumbnail==12.7.0
sqlparse==0.4.1
